requirement.txt
astroid==2.3.1
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-authority==0.13.2
django-silk==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
et-xmlfile==1.0.1
ffmpy==0.2.2
flake8==3.7.8
gprof2dot==2019.11.30
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jdcal==1.4.1
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.3
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
openpyxl==2.6.4
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
pycodestyle==2.5.0
pyfcm==1.4.7
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.2
pylint-django==2.0.13
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
twilio==6.31.1
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
